Compatibility with constellations plugin
Fixes compatibility with the endless sky constellations plugin. You can now leave and return to the pug galaxy if you chose neutrality. If you did not choose neutrality, the wormhole still goes to Queri. This way the plugin is kept intact as much as possible. This does cause potential lore issues, but when combining plugins you should expect to have issues anyways and this is better than being essentially softlocked.