Expand lore for Han Sizer month
On PR 7805 by Zitchas, I mistakenly thought on stopover appeared on every stopover. There was going to be a popup that appeared on the last stopover of a Han Sizer month job, but it was removed as it was thought to be annoying. I have readded it to the longest, rarest mission, so that you at least get something after having visited four planets. In vanilla, you would visit four planets, get absolutely nothing, then finally get a reward when landing on the destination. This helps add more lore for the Free Worlds, which is part of the main plot, making you feel more for them.