Add road to hai reveal danforth
Add the first mission in road to hai reveal, introducing danforth and making sure you can handle intense combat situations. Requires you to have met the Hai on your own. Also allows the Hai wormhole warning to happen after the main plot, even if you met the Hai in the middle of it.