Improve UX when disabling plugin
Fleet changes are now additive only, so they can be undone when uninstalling or disabling the plugin. Replaced Mark II ships in fleets with plugin specific versions so they are removed when uninstalling or disabling the plugin. Fixed a corner case in checkmate that could lead to you being locked out of Free Worlds when disabling or uninstalling the plugin. Removed the pug wormhole patch and added a mission to unpatch it in old saves as it caused issues when disabling or uninstalling the plugin and Constellations no longer breaks the wormhole.